Render needs to be a function.
Also I took out `screen.render();` from `ui/window.js` because
I think that the render function should always be called when
it's ready to be called. For now, that's still at start up,
but it won't always be. We'll want to load configuration
files and configure everything and have everything
ready before we render the screen.