Updating channel function width and adding more examples in index
channel.width is now a variable that holds the width of the
userlist so we don't have to call the reduce function all the
time. The function width() has been renamed to _width() and
should only be called when we need to figure out what the new
width of the userlist is.